invalidate() allows the session to be used again
After a call to invalidate() the session should be cleared, but ready
for use. At this point we reset the created timestamp, set the session
to new again, and generate a new session id.

The plug is told to clear the old session information first, so that in
the case of a session fixation attack the old session can not be reused.